JFS: detect and fix invalid directory index values
The directory index values are the unique cookies used to resume
a readdir at the proper place.  These are stored with each entry
in a directory.  fsck.jfs does not currently validate these entries,
nor even create them when populating the lost+found directory.

This patch causes readdir to detect the invalid cookies, and generate
new ones, if possible.