kbuild: Make scripts/Configure follow the definition of 'int'
Currently, scripts/Configure has code for the 'int' verb to take a
min/max.  This violates the spec described in
Documentation/kbuild/config-language.txt.  It also requires that if a
default is outside of +/- 10,000,000 that defaults be provided, or
'config' and 'oldconfig' will get stuck.  The following removes the
support for a min/max from scripts/Configure.

(by Tom Rini)