[PATCH] /proc/meminfo alterations for hugetlbpages
The patch from Rohit and David M-T changes the hugetlb page info in
/proc/meminfo slightly.

It makes the identifiers a little clearer while ensuring that we don't
add any identifiers which have whitespace.  glibc is/shall be parsing
this information to determine the size and alignment requirements of
the hugetlb pages.

This basically means that procfs is a requirement for successful
hugetlb page usage.  Not very nice, but I suspect real-world userspace
fails without procfs anyway.