[PATCH] direct-io bio_add_page fix
Patch from Badari Pulavarty

There was a corner case in the conversion of direct-io to use
bio_add_page() where we would start a new page out at the wrong sector
number.

Fix that by explicitly passing in the current page's starting sector,
and use that in the new BIO if we have to open a new one.

Fix an error-path page->count leak in dio_bio_add_page().