[PATCH] Forward port of 2.4 fsync_buffers_list() fix.
there was a bug in fysnc_buffers_list() in 2.4 (fixed in 2.4.19) that
could cause the function to return without having written the current
contents of all the buffers.

Obviously, this could be bad for anybody relying on ordering using
O_SYNC or fsync().  If an I/O was already in flight for a particular bh
at the time of the call to fsync_buffers_list(), ll_rw_block() will not
initiate a new I/O even though the contents may have changed.  It is
therefore necessary to wait before the call.  Here's a patch against
2.5.42 that applies the same fix.