kbuild: Allow for <mod>-y as well as <mod>-objs for multipart objects.
Traditionally, the individual components of a multipart module are
listed in <mod>-objs. Allow for using <mod>-y as well, as that turns
out to simplify declaring optional parts of multi-part modules, see
the converted examples in net/*/Makefile.

This change is backwards-compatible, i.e. not converted Makefiles
still work just fine.